2017 day13: check scanners without running them
This changes part two to not bother updating each scanner to where they
need to be and just calculates what we need to check. This made it
clearer for me to see where I could put in a break to speed things up.
Makes things way faster than they were.

This also changes the Scanner class to be the one I originally used from
part 2.