2018 day11: Speed up by calculating the sides of each square
Since I already had a matrix with the sum of each square, I realised I
could get the next one by droping the last column and row and adding an
extra row and column to each internal square.
That sped things up a bit, but I was summing that extra row and column
each time and I knew that wasn't necessary. After much pain with index
values I changed it so that only the first and last values of the extra
sides need to be changed each iteration. It was nice after this change
to see it sped up after every iteration.