2018 Day 19: Add solution for day 19
This was similar to day 23 from 2017 which I did the other day. I
tried the same approach, converting the program to C, from which I
worked out it was calculating the sum of the divisors that were stored
in one of the registers. After it starts that part it doesn't go back to
the other part of the program.

For the sum of the divisors of a number we only need to consider the
numbers up to the sqrt of the number when looking for divisors, so using
that shortcut I jumped out of the loop when we reached that part of the
program and calculated that more efficiently.

Similar to day 23 from last year, without seeing more input, I don't
know how a generic solution could be made.