2019 day 2: Solved by hand (and calculator)
The avoided using brute force for part 2.
The input happened to result in a linear Diophantine equation. It was
then easy to find suitable numbers that work for part 2.

I suppose it would be possible for the input to result in a non-linear
equation as two of the starting terms could be multiplied together, so
not sure I can create a general solution from this.