2019 day 7: fix interpreter to run with test data
I was passing the same copy of the input into all the interpreters. That
caused problems since the programs changed while being run, causing
infinite loops on the test data, but not the actual input data.

That makes me wonder if the input was designed in a way for multiple
interpreters to use it in parallel.