Move Python solutions for 2019 under python directory
This should allow me to add solutions in other languages more easily.
This also gave me an opportunity to use modules as I'd started to do so,
but I wasn't fully commited, so kept the directories named as numbers which
is, of course, invalid for a python module (though `python -m` seems to
run them fine).

Maybe I'll get to cleaning previous years up someday.